Moved hierarchicalexec code to use in other files.
